Dll completed sort of
Finished up the majority of the programming for this dll part. It is now a dll instead of exe. It also uses directx function calls and setup for buffers in direct x and stuff when loading the OBJ parsed data. But there is an error in the main code, will have to figure out how to diagnose through here. Though I imagine it has something to do with the main code which I discussed in it's commit.